Close Stream during Refresh
This change closes the stream during refresh.
This problem can best be seen on Firefox and the basic demo. If you
click refresh the camera is lost for ever and the browser needs to be
reloaded completely. This solves this issues by maintaining a reference
to the stream and unloading it (stop) during refresh. The camera can
then be used on the next cycle.